% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\docType{data}
\name{a_woman_of_no_importance}
\alias{a_woman_of_no_importance}
\title{The text of Oscar Wilde's play "A Woman of No Importance" (1893)}
\format{A character vector with 3274 elements}
\source{
\url{http://www.gutenberg.org/ebooks/854}
}
\usage{
a_woman_of_no_importance
}
\description{
A dataset containing the text of Oscar Wilde's 1893 play A Woman of "No 
Importance".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
