% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\docType{data}
\name{an_ideal_husband}
\alias{an_ideal_husband}
\title{The text of Oscar Wilde's play "An Ideal Husband" (1895)}
\format{A character vector with 4464 elements}
\source{
\url{http://www.gutenberg.org/ebooks/885}
}
\usage{
an_ideal_husband
}
\description{
A dataset containing the text of Oscar Wilde's 1893 play ""An Ideal Husband".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
