% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austen_books.R
\name{austen_books}
\alias{austen_books}
\title{Tidy data frame of Jane Austen's 6 completed, published novels}
\usage{
austen_books()
}
\value{
A data frame with two columns: text and book
}
\description{
Returns a tidy data frame of Jane Austen's 6 completed, published novels with 
two columns: \code{text}, which contains the text of the novels divided into 
elements of up to about 70 characters each, and \code{book}, which contains the titles of
the novels as a factor in order of publication.
}

