% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{emma}
\alias{emma}
\title{The text of Jane Austen's novel "Emma"}
\format{A character vector with 13683 elements}
\source{
\url{http://www.gutenberg.org/ebooks/158}
}
\usage{
emma
}
\description{
A dataset containing the text of Jane Austen's 1815 novel "Emma".
The UTF-8 plain text was sourced from the Gutenberg Project
and is divided into elements of about 70 characters each.
}
\keyword{datasets}

