% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janeaustenr-package.R
\docType{package}
\name{janeaustenr-package}
\alias{janeaustenr}
\alias{janeaustenr-package}
\title{janeaustenr: Jane Austen's Complete Novels}
\description{
Full texts for Jane Austen's 6 completed novels, ready for text analysis. These novels are "Sense and Sensibility", "Pride and Prejudice", "Mansfield Park", "Emma", "Northanger Abbey", and "Persuasion".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/juliasilge/janeaustenr}
  \item Report bugs at \url{https://github.com/juliasilge/janeaustenr/issues}
}

}
\author{
\strong{Maintainer}: Julia Silge \email{julia.silge@gmail.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})

}
\keyword{internal}
