% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{package}
\name{janeaustenr}
\alias{janeaustenr}
\alias{janeaustenr-package}
\title{An R Package for Jane Austen's Complete Novels}
\description{
This package contains the complete text of Jane Austen's 6 completed, 
published novels, formatted to be convenient for text analysis.
}

