% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\docType{data}
\name{lady_windermeres_fan}
\alias{lady_windermeres_fan}
\title{The text of Oscar Wilde's play "Lady Windermere's Fan" (1892)}
\format{A character vector with 2912 elements}
\source{
\url{http://www.gutenberg.org/ebooks/790}
}
\usage{
lady_windermeres_fan
}
\description{
A dataset containing the text of Oscar Wilde's 1892 play "Lady Windermere's 
Fan".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
