% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{northangerabbey}
\alias{northangerabbey}
\title{The text of Jane Austen's novel "Northanger Abbey"}
\format{A character vector with 7840 elements}
\source{
\url{http://www.gutenberg.org/ebooks/121}
}
\usage{
northangerabbey
}
\description{
A dataset containing the text of Jane Austen's novel "Northanger Abbey", 
published posthumously in 1818.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}

