% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\docType{package}
\name{oscarwilder}
\alias{oscarwilder}
\alias{oscarwilder-package}
\title{Oscar Wilde's Complete Plays}
\description{
This package contains the complete text of Oscar Wilde's 7 completed, 
published plays, formatted to be convenient for text analysis.
}
