% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\docType{data}
\name{salome}
\alias{salome}
\title{The text of Oscar Wilde's play "Salome" (1891)}
\format{A character vector with 2522 elements}
\source{
\url{http://www.gutenberg.org/ebooks/42704}
}
\usage{
salome
}
\description{
A dataset containing the text of Oscar Wilde's 1891 play "Salome".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
