% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{sensesensibility}
\alias{sensesensibility}
\title{The text of Jane Austen's novel "Sense and Sensibility"}
\format{A character vector with 10596 elements}
\source{
\url{http://www.gutenberg.org/ebooks/161}
}
\usage{
sensesensibility
}
\description{
A dataset containing the text of Jane Austen's 1811 novel "Sense and 
Sensibility". The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of about 70 characters each.
}
\keyword{datasets}

