% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\docType{data}
\name{the_dutchess_of_padua}
\alias{the_dutchess_of_padua}
\title{The text of Oscar Wilde's play "The Dutchess of Padua" (1891)}
\format{A character vector with 5470 elements}
\source{
\url{http://www.gutenberg.org/ebooks/875}
}
\usage{
the_dutchess_of_padua
}
\description{
A dataset containing the text of Oscar Wilde's 1891 play "The Dutchess of
Padua".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
