% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\docType{data}
\name{the_importance_of_being_earnest}
\alias{the_importance_of_being_earnest}
\title{The text of Oscar Wilde's play "The Importance of Being Earnest" (1895)}
\format{A character vector with 3091 elements}
\source{
\url{http://www.gutenberg.org/ebooks/854}
}
\usage{
the_importance_of_being_earnest
}
\description{
A dataset containing the text of Oscar Wilde's 1893 play "The Importance of
Being Earnest".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
