% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\docType{data}
\name{vera_or_the_nihilists}
\alias{vera_or_the_nihilists}
\title{The text of Oscar Wilde's play "Vera; Or, The Nihilists" (1880)}
\format{A character vector with 2835 elements}
\source{
\url{http://www.gutenberg.org/ebooks/26494}
}
\usage{
vera_or_the_nihilists
}
\description{
A dataset containing the text of Oscar Wilde's 1880 play "Vera; Or, The
Nihilists".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
