% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilde_plays.R
\name{wilde_plays}
\alias{wilde_plays}
\title{Tidy data frame of Oscar Wilde's 7 completed, presented plays}
\usage{
wilde_plays()
}
\value{
A data frame with two columns: \code{text} and \code{play}
}
\description{
Returns a tidy data frame of Oscar Wilde's 7 completed, presented plays with 
two columns: \code{text}, which contains the text of the plays divided into 
elements of up to about 70 characters each, and \code{play}, which contains
the titles of the plays as a factor in order of publication.
}
\details{
Users should be aware that there are some differences in usage 
between the novels as made available by Project Gutenberg. For example, 
"anything" vs. "any thing", "Mr" vs. "Mr.", and using underscores vs. all 
caps to indicate italics/emphasis.
}
\examples{

library(dplyr)

wilde_plays() \%>\% 
    group_by(play) \%>\%
    summarise(total_lines = n())

}
