% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{emma}
\alias{emma}
\title{The text of Jane Austen's novel "Emma"}
\format{A character vector with 15297 elements}
\source{
\url{http://www.gutenberg.org/ebooks/158}
}
\usage{
emma
}
\description{
A dataset containing the text of Jane Austen's 1815 novel "Emma".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each. 
(Some elements are blank.)
}
\keyword{datasets}
