% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{persuasion}
\alias{persuasion}
\title{The text of Jane Austen's novel "Persuasion"}
\format{A character vector with 7210 elements}
\source{
\url{http://www.gutenberg.org/ebooks/105}
}
\usage{
persuasion
}
\description{
A dataset containing the text of Jane Austen's novel "Persuasion", 
published posthumously in 1818.
The UTF-8 plain text was sourced from the Gutenberg Project
and is divided into elements of about 70 characters each.
}
\keyword{datasets}

