% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{prideprejudice}
\alias{prideprejudice}
\title{The text of Jane Austen's novel "Pride and Prejudice"}
\format{A character vector with 10721 elements}
\source{
\url{http://www.gutenberg.org/ebooks/1342}
}
\usage{
prideprejudice
}
\description{
A dataset containing the text of Jane Austen's 1813 novel "Pride and 
Prejudice". The UTF-8 plain text was sourced from the Gutenberg Project
and is divided into elements of about 70 characters each.
}
\keyword{datasets}

